refactor (all) : domain 패키지 추가하고 파일들의 위치 재조정
