+++ src/main/java/lotto/domain/LottoRank.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.math.BigDecimal;
 import java.util.Arrays;
