+++ src/test/java/lotto/domain/LottoCheckTest.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.domain;
 
-import org.assertj.core.api.Assertions;
+import lotto.domain.Lotto;
+import lotto.domain.LottoCheck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -8,7 +9,7 @@ import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
 
-import static lotto.LottoRank.*;
+import static lotto.domain.LottoRank.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertAll;
